Update README.md
I spend 10-20 hours working on this project every week. Most of this time is spent googling how to do something, trying that and more often than not failing. This is due to my novice ruby/rails skills. This is a frustrating, time consuming, and defeating process.

 I would really appreciate seeing others making *any* progress on other user features or doing *anything* that improves the behavior or capabilities of the app. I am not here to play the blame game or cause a scene, but I honestly do not think I can complete this project by the end of the semester without significantly more investment from every person on this team. I do not have time to help every person get things done. We will fail at this rate.  (it is 8 April now)

Please familiarize yourself with the routes available, views available, controllers, models, the database schema. Open the app and click around. Review documentation on pivotal about how the database is supposed to be set up. Watch a youtube video on the concept of model view controller. Read the ruby on rails getting started guide. 

After this, talk to the scrum master, pick a user story, and go. Don’t push anything if you break the app or more tests fail after your changes.